refactor: what a tape can be is answered in byteutil
ValidateTapeSize lived in hosting/cli, on the argument that refusing belongs at
the boundary where a good message can be given. But the message is about the
language — a tape is one to thirty-two bytes wide — and the two answers to that
question were in different packages: byteutil.TapeSize clamped 64 to 32, cli
refused it, and the comment on each pointed at the other.

They are now side by side. Clamping is for a caller that needs a width to work
with; refusing is for one deciding whether to go on at all, and silently
clamping there would compile a program in a dialect nobody chose.

ResolveTapeSize stays in the host: precedence between a flag and a manifest is
not a question about tapes.
