+++ byteutil/tape_test.go
@@ -2,6 +2,7 @@ package byteutil
 
 import (
 	"bytes"
+	"strings"
 	"testing"
 
 	"github.com/holiman/uint256"
@@ -24,6 +25,41 @@ func TestTapeSizeNormalizes(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			if got := TapeSize(tc.in); got != tc.want {
 				t.Errorf("TapeSize(%d) = %d, want %d", tc.in, got, tc.want)
+			}
+		})
+	}
+}
+
+// Normalizing and refusing are the two answers to a width, and they differ where it matters:
+// 64 normalizes to 32 and is refused, because a program compiled at a width nobody chose is
+// worse than one that does not compile.
+func TestValidateTapeSize(t *testing.T) {
+	cases := []struct {
+		name    string
+		size    int
+		wantErr bool
+	}{
+		{name: "unset is fine", size: 0},
+		{name: "floor", size: 1},
+		{name: "default", size: 8},
+		{name: "ceiling", size: 32},
+		{name: "below the floor", size: -1, wantErr: true},
+		{name: "above the ceiling", size: 33, wantErr: true},
+	}
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *testing.T) {
+			err := ValidateTapeSize(tc.size)
+			if tc.wantErr {
+				if err == nil {
+					t.Fatal("expected an error")
+				}
+				if !strings.Contains(err.Error(), "between 1 and 32") {
+					t.Errorf("error = %q, want it to state the range", err)
+				}
+				return
+			}
+			if err != nil {
+				t.Errorf("unexpected error: %v", err)
 			}
 		})
 	}
